backend/requirements.txt
fastapi
uvicorn
websockets
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
huggingface_hub